update(test): tests chats friends mixed
